Add method to `JobStatus` class for retrieving an instance HRID
Fixes #12

This commit adds a new method, `JobStatus#instance_hrid`, that allows consumers to request retrieval of the instance HRID associated with the current job status as long as the status has already successfully completed. If not, a `Failure()` is returned.

Because the instance HRID is not immediately available, this commit also includes disentangling the ability to wait on an operation and the `JobStatus#status` method.

This commit has been tested against a running Folio instance and confirmed to work as expected.