Change the internal structure of `FolioClient#data_import`
This commit is an internal refactoring of how the `FolioClient` singleton interacts with the `DataImport` class. Now the client instantiates `DataImport` passing in only the client singleton, deferring the other parameters to the `DataImport#import` method. This brings a consistent approach to the singleton instance methods in the client and makes clearer what pattern we're using going forward.