Do not authenticate as part of client configuration
Related to sul-dlss/globus_client#107

I could not wrap my brain about getting this to work without refactoring the client to be more like GlobusClient. The main difference was that FolioClient prior to this commit had functions re-using the same HTTP methods to do heavy lifting, which was making it difficult for me to reason about what was happening and why. This is the second time I've struggled to make sense of it (see: https://github.com/sul-dlss/folio_client/pull/27#discussion_r1122428121), so I went all in on the refactor which brings parity to our client gems. I see the main trade-off as now we have more code duplication and less swirly dependencies?

For the initial token, use a dummy value to avoid hitting any APIs during configuration, allowing the TokenWrapper to handle auto-magic token refreshing. Why not immediately get a valid token? Our apps commonly invoke client .configure methods in the initializer in all application environments, even those that are never expected to connect to production APIs, such as local development machines.

NOTE: nil and blank string cannot be used as dummy values here as they lead to a malformed request to be sent, which triggers an exception not rescued by TokenWrapper