instead of eagerly fetching a token on client initialization, lazily wait for the first request to fail, and use the token refresh mechanism
this is a recreation of @mjgiarlo's https://github.com/sul-dlss/folio_client/pull/77 minus the refactoring and TokenWrapper.refresh behavioral change.  thus, it preserves the current behavior of only re-executing the one failed request after re-authenticating, instead of it and all prior API calls made in the given high-level FolioClient method (e.g. FolioClient.fetch_hrid).

we'll iterate in subsequent commits on making the token refresh mechanism easier to understand at a glance.  but this fixes the issue of e.g. bringing up a rails console in a dev env with no credentials configured, and a possible honeybadger alert being logged.